webserver/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tortoise-orm==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.9.1
